Add script that check the Capabilities of WFS services
If the previous check (HTTP status code) returns code 200, then we scan the services URLs looking from the ones that obviously WFS and check them for the all appropriate versions of the protocol (1.0. and/or 1.1.0 and/or 2.0.0). We store the GetCapabilities answer, heck if it contains error messages, if is returned at all, etc.